Use same data type for resampling ratio as libsamplerate.

+++ src/resampler.cpp
@@ -20,7 +20,7 @@
 #include "resampler.hpp"
 
 Resampler::Resampler(const int& input_rate, const int& output_rate) : output(1, 0) {
-  resample_ratio = static_cast<float>(output_rate) / static_cast<float>(input_rate);
+  resample_ratio = static_cast<double>(output_rate) / static_cast<double>(input_rate);
 
   src_state = src_new(SRC_SINC_FASTEST, 1, nullptr);
 }
